+++ urssu/src/main/kotlin/com/example/urssu/controller/UserController.kt
@@ -4,10 +4,9 @@ import com.example.urssu.config.BaseException
 import com.example.urssu.config.BaseResponse
 import com.example.urssu.config.BaseResponseStatus
 import com.example.urssu.domain.entity.UserEntity
-import com.example.urssu.domain.repository.UserRepository
-import com.example.urssu.dto.JoinReqUserDto
-import com.example.urssu.dto.JoinResUserDto
-import com.example.urssu.dto.UserInfoDto
+import com.example.urssu.dto.user.JoinReqUserDto
+import com.example.urssu.dto.user.JoinResUserDto
+import com.example.urssu.dto.user.UserInfoDto
 import com.example.urssu.service.UserService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.validation.BindingResult
