+++ urssu/src/main/kotlin/com/example/urssu/dto/article/ArticleReqDto.kt
@@ -1,22 +1,24 @@
-package com.example.urssu.dto
+package com.example.urssu.dto.article
 
 import com.example.urssu.domain.entity.ArticleEntity
-import com.example.urssu.domain.entity.CommentEntity
 import com.example.urssu.domain.entity.UserEntity
 import javax.validation.constraints.NotBlank
 
-data class CommentReqDto (
+data class ArticleReqDto (
     val email: String,
 
     val password: String,
 
     @field: NotBlank
+    val title: String,
+
+    @field: NotBlank
     val content: String
 ){
-    fun toEntity(userEntity: UserEntity, articleEntity: ArticleEntity): CommentEntity {
-        return CommentEntity(
+    fun toEntity(userEntity: UserEntity): ArticleEntity{
+        return ArticleEntity(
+            title = title,
             content = content,
-            articleEntity = articleEntity,
             userEntity = userEntity
         )
     }
