+++ urssu/src/main/kotlin/com/example/urssu/dto/comment/CommentResDto.kt
@@ -1,4 +1,4 @@
-package com.example.urssu.dto
+package com.example.urssu.dto.comment
 
 data class CommentResDto(
     val commentId: Int,
