+++ urssu/src/main/kotlin/com/example/urssu/dto/user/LoginReqUserDto.kt
@@ -1,0 +1,6 @@
+package com.example.urssu.dto.user
+
+data class LoginReqUserDto (
+    val email: String,
+    val password: String
+)
