+++ urssu/src/main/kotlin/com/example/urssu/service/UserService.kt
@@ -1,7 +1,6 @@
 package com.example.urssu.service
 
 import com.example.urssu.config.BaseException
-import com.example.urssu.config.BaseResponse
 import com.example.urssu.config.BaseResponseStatus
 import com.example.urssu.domain.entity.ArticleEntity
 import com.example.urssu.domain.entity.CommentEntity
@@ -9,8 +8,8 @@ import com.example.urssu.domain.entity.UserEntity
 import com.example.urssu.domain.repository.ArticleRepository
 import com.example.urssu.domain.repository.CommentRepository
 import com.example.urssu.domain.repository.UserRepository
-import com.example.urssu.dto.JoinReqUserDto
-import com.example.urssu.dto.UserInfoDto
+import com.example.urssu.dto.user.JoinReqUserDto
+import com.example.urssu.dto.user.UserInfoDto
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
